Move App.js-related components to Shared folder
