+++ src/App.js
@@ -4,13 +4,13 @@ import { Provider } from 'react-redux';
 import { I18nextProvider } from 'react-i18next';
 import connect from 'redux-connect-decorator';
 import Router from './components/screens/router';
-import Loading from './components/loading';
+import Loading from './components/shared/loading';
 import store from './store/index';
 import ThemeContext from './contexts/theme';
 import { themes } from './constants/styleGuide';
 import i18n from '../locales';
-import Alert from './components/alert';
-import Modal from './components/modal';
+import Alert from './components/shared/alert';
+import Modal from './components/shared/modal';
 
 @connect(state => ({
   settings: state.settings,
