+++ src/components/shared/bookmarks/deleteBookmarkModal/styles.js
@@ -1,5 +1,5 @@
-import { themes, colors, fonts, boxes } from '../../../constants/styleGuide';
-import { deviceWidth } from '../../../utilities/device';
+import { themes, colors, fonts, boxes } from '../../../../constants/styleGuide';
+import { deviceWidth } from '../../../../utilities/device';
 
 export default () => ({
   common: {
