+++ src/components/shared/bookmarks/list.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 import { DraggableItem, Item } from './item';
-import withTheme from '../shared/withTheme';
+import withTheme from '../withTheme';
 import getStyles from './styles';
 
 const List = ({
