+++ src/components/shared/loading/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import connect from 'redux-connect-decorator';
 import { View } from 'react-native';
 import LottieView from 'lottie-react-native';
-import progressBar from '../../assets/animations/progressBar.json';
-import withTheme from '../shared/withTheme';
+import progressBar from '../../../assets/animations/progressBar.json';
+import withTheme from '../withTheme';
 import getStyles from './styles';
 
 @connect(state => ({
