+++ src/components/shared/modal/index.js
@@ -2,13 +2,13 @@ import React from 'react';
 import { View, ScrollView } from 'react-native';
 import ModalBox from 'react-native-modalbox';
 import { translate } from 'react-i18next';
-import { B } from '../shared/toolBox/typography';
-import { boxes } from '../../constants/styleGuide';
-import withTheme from '../shared/withTheme';
-import HeaderBackButton from '../screens/router/headerBackButton';
-import { deviceHeight, headerHeight } from '../../utilities/device';
+import { B } from '../toolBox/typography';
+import { boxes } from '../../../constants/styleGuide';
+import withTheme from '../withTheme';
+import HeaderBackButton from '../../screens/router/headerBackButton';
+import { deviceHeight, headerHeight } from '../../../utilities/device';
 import getStyles from './styles';
-import ModalHolder from '../../utilities/modal';
+import ModalHolder from '../../../utilities/modal';
 
 class Modal extends React.Component {
   state = {
