+++ src/components/shared/modal/styles.js
@@ -1,7 +1,7 @@
 import { Platform } from 'react-native';
-import { themes, colors, boxes } from '../../constants/styleGuide';
-import { setColorOpacity } from '../../utilities/helpers';
-import { deviceType } from '../../utilities/device';
+import { themes, colors, boxes } from '../../../constants/styleGuide';
+import { setColorOpacity } from '../../../utilities/helpers';
+import { deviceType } from '../../../utilities/device';
 
 export default () => ({
   common: {
